caching: Split caches into new files.
